Fix up semantics of abstractions and traits
