+++ plugins/woocommerce/src/Admin/Features/ProductBlockEditor/ProductTemplates/BaseProductTemplate.php
@@ -1,173 +1,86 @@
 <?php
 /**
- * Product Block Editor abstract base product template.
+ * Product Block Editor reusable template pieces.
  */
 
 namespace Automattic\WooCommerce\Admin\Features\ProductBlockEditor\ProductTemplates;
 
-abstract class BaseProductTemplate {
-    use CoreProductTemplate;
+trait BaseProductTemplate {
 
     /**
-     * Block tree root.
+     * Add all base template groups.
      */
-    const ROOT = 'root';
+    protected function add_base_template() {
+        $this->add_general_group();
+        $this->add_pricing_group();
+    }
 
     /**
-     * Index for the block properties.
+     * Add general group and fields.
      */
-    const BLOCK_PROPERTIES_INDEX = 1;
-
-    /**
-     * Index for the child blocks.
-     */
-    const BLOCK_CHILDREN_INDEX = 2;
-
-    /**
-     * Property for the internal order.
-     */
-    const ORDER_PROPERTY = '_order';
-
-    /**
-     * The block template.
-     */
-    protected $template = array();
-
-    /**
-     * Cache of reference to inserted blocks location in tree.
-     */
-    protected $cache = array();
-
-    /**
-     * Set up the template.
-     */
-    public function __construct() {
-        $this->add_group( array(
-            'id'    => 'general2',
-            'title' => __( 'General2', 'woocommerce' ),
-            'order' => 20,
-        ) );
-        $this->add_group( $this->get_general_tab_args() );
-
-        $this->add_section( $this->get_basic_section_args() );
-        $this->add_field(
+    protected function add_general_group() {
+        $this->add_group(
             array(
-                'parent' => 'basic-details',
-                'block'  => $this->get_name_field_args(),
-                'order'  => '20',
+                'id'    => self::GENERAL_GROUP,
+                'title' => __( 'General', 'woocommerce' ),
+                'order' => 10,
+            )
+        );
+        $this->add_section(
+            array(
+                'id'          => self::BASIC_DETAILS_SECTION,
+                'title'       => __( 'Basic details', 'woocommerce' ),
+                'description' => __( 'This info will be displayed on the product page, category pages, social media, and search results.', 'woocommerce' ),
+                'parent'      => self::GENERAL_GROUP,
             )
         );
         $this->add_field(
             array(
-                'parent' => 'basic-details',
-                'block'  => array( 'test-block' ),
-                'order'  => '10',
-            )
-        );
-    }
-
-    /**
-     * Add a field to the existing template.
-     */
-    protected function add_field( $field ) {
-        $args   = wp_parse_args( $field, array( 'order' => 10 ) );
-        $id     = $args['id'] ?? null;
-        $block  = $args['block'];
-        $parent = $args['parent'];
-        $order  = $args['order'];
-
-        $block[ self::BLOCK_PROPERTIES_INDEX ][ self::ORDER_PROPERTY ] = $order;
-
-        $this->insert_block( $parent, $block, $id );
-    }
-
-    /**
-     * Insert a block into the template tree.
-     */
-    private function insert_block( $parent, $block, $id ) {
-        if ( $parent === self::ROOT ) {
-            $blocks = &$this->template;
-        } else {
-            $blocks = &$this->cache[ $parent ][ self::BLOCK_CHILDREN_INDEX ];
-        }
-
-        $blocks[] = $block;
-
-        if ( $id ) {
-            $index = count( $blocks ) - 1;
-            $this->cache[ $id ] = &$blocks[ $index ];
-        }
-    }
-
-    /**
-     * Add a group to the template.
-     */
-    protected function add_group( $args = array() ) {
-        $args = wp_parse_args( $args, array( 'order' => 10 ) );
-
-        $group = array(
-            'woocommerce/product-tab',
-            array(
-                'id'    => $args['id'],
-                'title' => $args['title'],
-                'order' => $args['order'],
+                'parent' => self::GENERAL_GROUP,
+                'block'  => array(
+                    'woocommerce/product-name-field',
+                    array(
+                        'name' => 'Product name',
+                    ),
+                ),
             ),
-            array(),
         );
-
         $this->add_field(
             array(
-                'id'     => $args['id'],
-                'parent' => self::ROOT,
-                'block'  => $group,
-                'order'  => $args['order'] ?? null,
-            )
-        );
-    }
-
-    /**
-     * Add a section to the template.
-     */
-    protected function add_section( $args = array() ) {
-        $section = array(
-            'woocommerce/product-section',
-            array(
-                'id'          => $args['id'],
-                'title'       => $args['title'],
-                'description' => $args['description'],
+                'parent' => self::GENERAL_GROUP,
+                'block'  => array(
+                    array(
+                        'woocommerce/product-summary-field',
+                    ),
+                ),
             ),
-            array(),
         );
+    }
 
-        $this->add_field(
+    /**
+     * Add pricing group and fields.
+     */
+    protected function add_pricing_group() {
+        $this->add_group(
             array(
-                'id'     => $args['id'],
-                'parent' => $args['parent'],
-                'block'  => $section,
-                'order'  => $args['order'] ?? null,
+                'id'    => self::PRICING_GROUP,
+                'title' => __( 'Pricing', 'woocommerce' ),
+                'order' => 20,
+            )
+        );
+        $this->add_section(
+            array(
+                'id'          => self::PRICING_SECTION,
+                'title'       => __( 'Pricing', 'woocommerce' ),
+                'description' => sprintf(
+                    /* translators: %1$s: Images guide link opening tag. %2$s: Images guide link closing tag.*/
+                    __( 'Set a competitive price, put the product on sale, and manage tax calculations. %1$sHow to price your product?%2$s', 'woocommerce' ),
+                    '<a href="https://woocommerce.com/posts/how-to-price-products-strategies-expert-tips/" target="_blank" rel="noreferrer">',
+                    '</a>'
+                ),
+                'parent'      => self::PRICING_GROUP,
             )
         );
     }
 
-    /**
-     * Sort blocks recursively by the internal block attribute order property.
-     */
-    private function sort_by_order( $blocks ) {
-        usort( $blocks, function( $a, $b ) {
-            return $a[ self::BLOCK_PROPERTIES_INDEX ][ self::ORDER_PROPERTY ] > $b[ self::BLOCK_PROPERTIES_INDEX ][ self::ORDER_PROPERTY ] ? 1 : -1;
-        } );
-        foreach( $blocks as $index => $block ) {
-            if ( isset( $block[ self::BLOCK_CHILDREN_INDEX ] ) ) {
-                $blocks[ $index ][ self::BLOCK_CHILDREN_INDEX ] = $this->sort_by_order( $block[ self::BLOCK_CHILDREN_INDEX ] );
-            }
-        }
-        return $blocks;
-    }
-
-    /**
-     * Get the template.
-     */
-    public function get_template() {
-        return $this->sort_by_order( $this->template );
-    }
 }
